Add GitHub action and container (#2)
I thought it would be fun to add a container to serve two purposes:

 - run contrib using a Docker container (and more importantly)
 - creating a GitHub action for it, deployed at the repository.

To make this work, I did the following, represented in this PR:

 - added a Dockerfile to generate a container with config installed and a custom entrypoint
  - the entrypoint.sh is expecting the GitHub action envars, but I show the user how to use the container with a different entrypoint (to contrib directly) or using the same action envars.
  - the action.yml file that defines inputs and outputs to drive the generation along with metadata for GitHub if we release on the marketplace
 - verbose instructions in the README for building and using the docker container, and writing a yaml flie to put in .github/workflows
 - a workflow in .github/workflows that runs the action from here, and uploads a pdf as artifact.

I chose settings that would run fairly quickly (e.g., 2 commits and only s* prefix files in one of the package folders) and also added fairly verbose printing of the variables in case the action needs to be debugged. 

![image](https://user-images.githubusercontent.com/814322/82741005-2c339300-9d0b-11ea-903d-4f01452ca4e9.png)

For those settings I've tested a few times, and it takes between 6 and 7 minutes, including the container build. The example also sets a cache, so I suspect that would speed it up if the cache is used in a merged PR. You can see a full run [here](https://github.com/researchapps/contrib/runs/702697927?check_suite_focus=true) at a PR to my own repository, and example pdfs generated are here:  
 
[loc-in-packages-by-org.pdf](https://github.com/spack/contrib/files/4672560/loc-in-packages-by-org.pdf)
[loc-in-packages-by-author.pdf](https://github.com/spack/contrib/files/4672561/loc-in-packages-by-author.pdf)

I hope this is useful! I thought it was really cool, and I'd want to use it, but I don't want it to be too hard in terms of needing to remember to do it :)

Signed-off-by: vsoch <vsochat@stanford.edu>